Fix/clean-up the build for lmbench
This is the heads-up that meta-oe of Yocto is planning on integrating a couple
of changes to lmbench.

These changes can be wrapped up around fixing/cleaning up as follows:
- if lmbench doesn't use a return from a function attributed to warn_unused_result
  supress the warning/s
- fix buffer overflow
- define feature_test_macro _GNU_SOURCE compiling in sched_{get,set}affinity()
- go strict for a pre-build check if sched_{get,set}affinity() from glibc builds
  cleanly. If it is not compile it out.

Please find the original discussion on the changes at meta-oe
http://lists.openembedded.org/pipermail/openembedded-devel/2019-October/202206.html

I have tested these changes against this repo and they build just fine.
The compiler warns though about the changes which already are in the meta-oe and
are not part of this commit.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>
Reviewed-off-by: Marcin Lapaj <marcin.lapaj@tieto.com>